Added files via upload
These are two videos which playback the spike generation corresponding to pixels rate. You can see pixels with more density (the righter pixels) have more spikes/sec. I assumed 100ms for pixels representation. In fact I used the pixels intensity as the intensity of stimuli which are clamped on the inputs of our SNN during of simulation. The maximum rate for completely white pixels is 1000 Hz (100 spikes in 100 ms).